refactor: `schemas.py` split in `schemas/*.py`
Additionally, I've also rewritten the docstrings to make those clearer and following a more standardized format
