+++ src/argilla/client/feedback/schemas/__init__.py
@@ -1,0 +1,49 @@
+#  Copyright 2021-present, the Recognai S.L. team.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+from argilla.client.feedback.schemas.fields import FieldSchema, FieldTypes, TextField
+from argilla.client.feedback.schemas.questions import (
+    LabelQuestion,
+    MultiLabelQuestion,
+    QuestionSchema,
+    QuestionTypes,
+    RankingQuestion,
+    RatingQuestion,
+    TextQuestion,
+)
+from argilla.client.feedback.schemas.records import (
+    FeedbackRecord,
+    RankingValueSchema,
+    ResponseSchema,
+    SuggestionSchema,
+    ValueSchema,
+)
+
+__all__ = [
+    "RatingQuestion",
+    "TextQuestion",
+    "LabelQuestion",
+    "MultiLabelQuestion",
+    "RankingQuestion",
+    "QuestionSchema",
+    "QuestionTypes",
+    "FeedbackRecord",
+    "SuggestionSchema",
+    "ResponseSchema",
+    "ValueSchema",
+    "RankingValueSchema",
+    "TextField",
+    "FieldSchema",
+    "FieldTypes",
+]
