+++ src/argilla/client/feedback/schemas/utils.py
@@ -1,0 +1,34 @@
+#  Copyright 2021-present, the Recognai S.L. team.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+from typing import TYPE_CHECKING, Dict, List
+
+if TYPE_CHECKING:
+    from argilla.client.feedback.schemas.questions import QuestionSchema
+
+
+class LabelMappingMixin:
+    """Mixin to provide label mapping functionality to a `QuestionSchema`."""
+
+    @property
+    def __all_labels__(self: "QuestionSchema") -> List[str]:
+        return [entry["value"] for entry in self.settings["options"]]
+
+    @property
+    def __label2id__(self) -> Dict[str, int]:
+        return {label: idx for idx, label in enumerate(self.__all_labels__)}
+
+    @property
+    def __id2label__(self) -> Dict[int, str]:
+        return {idx: label for idx, label in enumerate(self.__all_labels__)}
